Patches for Docker e2e behaviour
