+++ docker-compose.yml
@@ -9,16 +9,18 @@ services:
       - API_PORT=8081
       - WALLET_SERVICE_URI=http://wallet:0000
       - EXECUTION_SERVICE_URI=http://execution_service:9000
-      - CONTRACT_SERVER_LOWER_PORT_BOUND=8082
-      - CONTRACT_SERVER_UPPER_PORT_BOUND=8900
+      - CONTRACT_DIRECTORY=/application/bundles
       - REDIS_HOST=redis
       - REDIS_PORT=6379
       - OPERATION_MODE=distributed
-    # volumes:
+    volumes:
     #   - ./dist/server:/application/dist/server
+      - ./test/bundles/docker:/application/bundles
     ports:
       - 8081:8081
       - 8082:8082
+    depends_on:
+      - execution_service
   execution_service:
     build:
       context: .
