+++ features/contract_loading.feature
@@ -9,7 +9,7 @@ Feature: Contract Loading
 
     Examples:
       | address              | pk                | method | methodArguments                       |
-      | "plutusGuessingGame" | "uniquePublicKey" | "initialise"  | "{\"originatorPk\": \"myPublicKey\"}" |
+      | "plutusGuessingGame" | "uniquePublicKey" | "schema"  | "{}" |
 
   Scenario Outline: a contract is loaded only once, with subsequent requests ignored
