+++ src/server/infrastructure/engine_clients/plutus/PlutusEngineClient.ts
@@ -1,5 +1,6 @@
 import { EngineClient } from '../../../../core'
 import { NetworkInterface } from '../../../../lib'
+import { RetryPromise } from 'promise-exponential-retry'
 
 type Config = {
   executionEndpoint: string,
@@ -10,10 +11,14 @@ export function PlutusEngineClient (config: Config): EngineClient {
   const { executionEndpoint, networkInterface } = config
   return {
     name: 'plutus',
+    // As we load the contracts at boot time, the execution service
+    // may not yet be available, so we allow a connection buffer here
     async loadExecutable ({ contractAddress, executable }) {
-      return networkInterface.post(`${executionEndpoint}/loadSmartContract`,
-        { contractAddress, executable: executable.toString('base64') }
-      )
+      return RetryPromise.retryPromise('loadContract', () => {
+        return networkInterface.post(`${executionEndpoint}/loadSmartContract`,
+          { contractAddress, executable: executable.toString('base64') }
+        )
+      }, 10)
     },
     async unloadExecutable (contractAddress) {
       return networkInterface.post(`${executionEndpoint}/unloadSmartContract`, { contractAddress })
