+++ test/bundles/create_docker_tar/Dockerfile
@@ -1,0 +1,10 @@
+FROM node:10.16-alpine
+
+RUN mkdir /application
+COPY index.js /application
+COPY package.json /application
+
+WORKDIR /application
+RUN npm i --production
+
+CMD ["node", "index.js"]
