+++ test/bundles/create_docker_tar/README.md
@@ -1,0 +1,3 @@
+Manually edit the contents in index.js to create the required mock container, then run `npm run docker-tar`
+
+Copy the output.tar.gz to the required location
